requirements.txt
alpaca-py==0.29.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1